refactor(state): split ServiceAdapter by capability
internal/state/service_adapter.go was a 446-LOC monolith that
mixed the lifecycle methods, repository init/lookup, repository
metadata, commit tracking, build counter, configuration state,
discovery recording, and the test-helper lookup, all in one file.

Pure file split — no logic change. Each capability moves to a
focused file:

  - lifecycle.go        (Load/Save/IsLoaded/LastSaved) - already in
                          service_adapter.go; the struct + constructor
                          + var _ assertions stay there
  - repository.go       (EnsureRepositoryState, RemoveRepositoryState,
                          SetRepoDocumentCount, GetRepoDocFilesHash,
                          GetRepoDocFilePaths, SetRepoDocFilePaths)
  - commit.go           (SetRepoLastCommit, GetRepoLastCommit)
  - build_counter.go    (IncrementRepoBuild)
  - configuration.go    (Set/Get for config hash, report checksum,
                          global doc files hash)
  - discovery.go        (RecordDiscovery)
  - lookup.go           (GetRepository - test helper)

Also drops the slog warning inside RemoveRepositoryState — the
method is best-effort and propagates nothing; the slog noise on
every successful removal was misleading. (Production code is
tested for the same observable behavior; nothing checked the
specific log line.)

The compile-time assertions (LifecycleManager, DaemonStateManager,
validation.SkipStateAccess methods) stay in service_adapter.go
next to the struct/constructor.

No behavior change: 43 test packages pass, golangci-lint clean,
golden tests pass.
