+++ internal/state/service_adapter.go
@@ -2,11 +2,9 @@ package state
 
 import (
 	"context"
-	"log/slog"
 	"sync"
 	"time"
 
-	"git.home.luguber.info/inful/docbuilder/internal/foundation"
 	"git.home.luguber.info/inful/docbuilder/internal/foundation/errors"
 )
 
@@ -16,6 +14,18 @@ import (
 //
 // The adapter translates simple method signatures (no context, no Result types)
 // to the typed Store method signatures (context + Result types).
+//
+// The implementation is split across multiple files by capability:
+//   - lifecycle.go:        Load, Save, IsLoaded, LastSaved
+//   - repository.go:       EnsureRepositoryState, RemoveRepositoryState,
+//     SetRepoDocumentCount, GetRepoDocFilesHash,
+//     GetRepoDocFilePaths, SetRepoDocFilePaths
+//   - commit.go:           SetRepoLastCommit, GetRepoLastCommit
+//   - build_counter.go:    IncrementRepoBuild
+//   - configuration.go:    Set/Get* (config hash, report checksum,
+//     global doc files hash)
+//   - discovery.go:        RecordDiscovery
+//   - lookup.go:           GetRepository (test helper)
 type ServiceAdapter struct {
 	service *Service
 	mu      sync.RWMutex
@@ -85,343 +95,6 @@ func (a *ServiceAdapter) LastSaved() *time.Time {
 	a.mu.RLock()
 	defer a.mu.RUnlock()
 	return a.lastSaved
-}
-
-// --- RepositoryInitializer interface ---
-
-// EnsureRepositoryState creates a repository entry if it doesn't exist.
-// For compatibility with legacy code, empty branch defaults to "main".
-func (a *ServiceAdapter) EnsureRepositoryState(url, name, branch string) {
-	if url == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-
-	// Check if repository already exists
-	existing := store.GetByURL(ctx, url)
-	if existing.IsOk() {
-		if opt := existing.Unwrap(); opt.IsSome() {
-			return // Already exists
-		}
-	}
-
-	// Default branch for compatibility with legacy code that passes empty branch
-	if branch == "" {
-		branch = defaultBranchMain
-	}
-
-	// Create new repository
-	repo := &Repository{
-		URL:       url,
-		Name:      name,
-		Branch:    branch,
-		CreatedAt: time.Now(),
-		UpdatedAt: time.Now(),
-	}
-	_ = store.Create(ctx, repo) // Ignore error for interface compatibility
-}
-
-// RemoveRepositoryState removes a repository entry from persistent state.
-// It is used by daemon orchestration to reflect discovery removals.
-func (a *ServiceAdapter) RemoveRepositoryState(url string) {
-	if url == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	res := store.Delete(ctx, url)
-	if res.IsErr() {
-		slog.Warn("Failed to delete repository state",
-			slog.String("url", url),
-			slog.Any("error", res.UnwrapErr()))
-	}
-}
-
-// --- RepositoryMetadataWriter interface ---
-
-// SetRepoDocumentCount sets the document count for a repository.
-func (a *ServiceAdapter) SetRepoDocumentCount(url string, count int) {
-	if url == "" || count < 0 {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	_ = store.SetDocumentCount(ctx, url, count)
-}
-
-// SetRepoDocFilesHash sets the document files hash for a repository.
-func (a *ServiceAdapter) SetRepoDocFilesHash(url, hash string) {
-	if url == "" || hash == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	_ = store.SetDocFilesHash(ctx, url, hash)
-}
-
-// --- RepositoryMetadataReader interface ---
-
-// GetRepoDocFilesHash returns the document files hash for a repository.
-func (a *ServiceAdapter) GetRepoDocFilesHash(url string) string {
-	if url == "" {
-		return ""
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	result := store.GetByURL(ctx, url)
-	if result.IsErr() {
-		return ""
-	}
-	opt := result.Unwrap()
-	if opt.IsNone() {
-		return ""
-	}
-	hashOpt := opt.Unwrap().DocFilesHash
-	if hashOpt.IsNone() {
-		return ""
-	}
-	return hashOpt.Unwrap()
-}
-
-// GetRepoDocFilePaths returns the document file paths for a repository.
-func (a *ServiceAdapter) GetRepoDocFilePaths(url string) []string {
-	if url == "" {
-		return nil
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	result := store.GetByURL(ctx, url)
-	if result.IsErr() {
-		return nil
-	}
-	opt := result.Unwrap()
-	if opt.IsNone() {
-		return nil
-	}
-	// Return a copy to prevent mutation
-	paths := opt.Unwrap().DocFilePaths
-	if len(paths) == 0 {
-		return nil
-	}
-	cp := make([]string, len(paths))
-	copy(cp, paths)
-	return cp
-}
-
-// --- RepositoryMetadataStore interface (extends Reader + Writer) ---
-
-// SetRepoDocFilePaths sets the document file paths for a repository.
-func (a *ServiceAdapter) SetRepoDocFilePaths(url string, paths []string) {
-	if url == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	_ = store.SetDocFilePaths(ctx, url, paths)
-}
-
-// --- RepositoryCommitTracker interface ---
-
-// SetRepoLastCommit sets the last commit hash for a repository.
-func (a *ServiceAdapter) SetRepoLastCommit(url, name, branch, commit string) {
-	if url == "" || commit == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-
-	// Get existing repository to update
-	result := store.GetByURL(ctx, url)
-	if result.IsErr() {
-		return
-	}
-	opt := result.Unwrap()
-	if opt.IsNone() {
-		// Repository doesn't exist, create it first
-		a.EnsureRepositoryState(url, name, branch)
-	}
-
-	// Update the repository with the new commit
-	result = store.GetByURL(ctx, url)
-	if result.IsErr() || result.Unwrap().IsNone() {
-		return
-	}
-	repo := result.Unwrap().Unwrap()
-	repo.LastCommit = foundation.Some(commit)
-	repo.UpdatedAt = time.Now()
-	_ = store.Update(ctx, repo)
-}
-
-// GetRepoLastCommit returns the last commit hash for a repository.
-func (a *ServiceAdapter) GetRepoLastCommit(url string) string {
-	if url == "" {
-		return ""
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	result := store.GetByURL(ctx, url)
-	if result.IsErr() {
-		return ""
-	}
-	opt := result.Unwrap()
-	if opt.IsNone() {
-		return ""
-	}
-	commitOpt := opt.Unwrap().LastCommit
-	if commitOpt.IsNone() {
-		return ""
-	}
-	return commitOpt.Unwrap()
-}
-
-// --- RepositoryBuildCounter interface ---
-
-// IncrementRepoBuild increments build counters for a repository.
-func (a *ServiceAdapter) IncrementRepoBuild(url string, success bool) {
-	if url == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	_ = store.IncrementBuildCount(ctx, url, success)
-}
-
-// --- ConfigurationStateStore interface ---
-
-// SetLastConfigHash stores the last config hash.
-func (a *ServiceAdapter) SetLastConfigHash(hash string) {
-	if hash == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetConfigurationStore()
-	_ = store.Set(ctx, "last_config_hash", hash)
-}
-
-// GetLastConfigHash returns the last config hash.
-func (a *ServiceAdapter) GetLastConfigHash() string {
-	ctx := context.Background()
-	store := a.service.GetConfigurationStore()
-	result := store.Get(ctx, "last_config_hash")
-	if result.IsErr() {
-		return ""
-	}
-	opt := result.Unwrap()
-	if opt.IsNone() {
-		return ""
-	}
-	if s, ok := opt.Unwrap().(string); ok {
-		return s
-	}
-	return ""
-}
-
-// SetLastReportChecksum stores the last report checksum.
-func (a *ServiceAdapter) SetLastReportChecksum(sum string) {
-	if sum == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetConfigurationStore()
-	_ = store.Set(ctx, "last_report_checksum", sum)
-}
-
-// GetLastReportChecksum returns the last report checksum.
-func (a *ServiceAdapter) GetLastReportChecksum() string {
-	ctx := context.Background()
-	store := a.service.GetConfigurationStore()
-	result := store.Get(ctx, "last_report_checksum")
-	if result.IsErr() {
-		return ""
-	}
-	opt := result.Unwrap()
-	if opt.IsNone() {
-		return ""
-	}
-	if s, ok := opt.Unwrap().(string); ok {
-		return s
-	}
-	return ""
-}
-
-// SetLastGlobalDocFilesHash stores the global doc files hash.
-func (a *ServiceAdapter) SetLastGlobalDocFilesHash(hash string) {
-	if hash == "" {
-		return
-	}
-	ctx := context.Background()
-	store := a.service.GetConfigurationStore()
-	_ = store.Set(ctx, "last_global_doc_files_hash", hash)
-}
-
-// GetLastGlobalDocFilesHash returns the global doc files hash.
-func (a *ServiceAdapter) GetLastGlobalDocFilesHash() string {
-	ctx := context.Background()
-	store := a.service.GetConfigurationStore()
-	result := store.Get(ctx, "last_global_doc_files_hash")
-	if result.IsErr() {
-		return ""
-	}
-	opt := result.Unwrap()
-	if opt.IsNone() {
-		return ""
-	}
-	if s, ok := opt.Unwrap().(string); ok {
-		return s
-	}
-	return ""
-}
-
-// --- Additional methods used by Daemon ---
-
-// RecordDiscovery records a discovery operation for a repository.
-// This mimics the legacy StateManager.RecordDiscovery method.
-func (a *ServiceAdapter) RecordDiscovery(repoURL string, documentCount int) {
-	if repoURL == "" {
-		return
-	}
-	ctx := context.Background()
-
-	// Update statistics using RecordDiscovery method
-	statsStore := a.service.GetStatisticsStore()
-	_ = statsStore.RecordDiscovery(ctx, documentCount)
-
-	// Update repository state
-	repoStore := a.service.GetRepositoryStore()
-	result := repoStore.GetByURL(ctx, repoURL)
-	if result.IsOk() {
-		if opt := result.Unwrap(); opt.IsSome() {
-			repo := opt.Unwrap()
-			now := time.Now()
-			repo.LastDiscovery = foundation.Some(now)
-			repo.DocumentCount = documentCount
-			repo.UpdatedAt = now
-			_ = repoStore.Update(ctx, repo)
-		}
-	}
-}
-
-// --- Test helper methods ---
-
-// GetRepository returns the stored repository state for url, or nil if not
-// found. It exposes the internal *Repository type so callers (notably the
-// daemon integration tests) can inspect fields directly.
-func (a *ServiceAdapter) GetRepository(url string) *Repository {
-	if url == "" {
-		return nil
-	}
-	ctx := context.Background()
-	store := a.service.GetRepositoryStore()
-	result := store.GetByURL(ctx, url)
-	if result.IsErr() {
-		return nil
-	}
-	opt := result.Unwrap()
-	if opt.IsNone() {
-		return nil
-	}
-	return opt.Unwrap()
 }
 
 // Compile-time verification that ServiceAdapter implements the canonical
